Plugin: Add indentation rules
Shamelessly copies from bundled JavaScript plugin. As a bonus, it
supports arrays of object out of the box, thanks to the
indentSquareBrackets key. But on the downside, it de-indented `default`
keyword in QML object block context (property modifier), so a pure QML
preferences had to be duplicated with "case" and "default" removed.